+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ52MNYT3WXW967HDHBDZN0M"
+        "01KJ576JE8WJP2AAWG20RKZA50"
     )
     r"""The unique identifier of the data source"""