+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ52MNYP9SMC59HW2XRSGXG3"
+        "01KJ576JE47136870PTJKWWZE8"
     )
     r"""The unique identifier of the data source"""