+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ52MNYSX7XNWGW4CBVMQK3T"
+        "01KJ576JE7YCKJF5XGH47GR1QD"
     )
     r"""The unique identifier of the data source"""