+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ52MNYYJNS5F668RBCDVV4N"
+        "01KJ576JECF2X1RCHAHD6NF5A1"
     )
     r"""The unique identifier of the data source"""